Support references to passes
The GCC prototype now emits the pass tree once at the top of the
file, with the optimization records making references to the
pertinent passes.

To support this, it seemed simplest to introduce Python classes
generated from the JSON, rather than reusing the JSON internally.

This also required retaining the translation units, since the
passes are per-TU.